separated sidebar from navbar
